requirements.txt
PyQt6
BeautifulSoup4
google-generativeai
serpapi
google-search-results
pandas
matplotlib
scikit-learn
torch
transformers
accelerate
python-dotenv